requirements.txt
praw>=6.1.0
python-telegram-bot>=11.1.0